move createChain to internals
